refactor: move validation code from api into fx-core
